Adjust pdata function table entries sorting for AArch64
The .pdata section contains an array of function table entries that
are used for exception handling. The entries should be sorted by
begin address, which is usually the first 4 bytes RVA in the entry.
Entry sizes are different for x64 and AArch64.
This difference is addressed in this patch.

bfd/ChangeLog:

	* peXXigen.c (defined): New.
	(_bfd_XXi_final_link_postscript): Use
	FUNCTION_TABLE_ENTRY_SIZE.